team-3301/requirements.txt
PyMuPDF==1.23.8
Pillow==10.1.0
requests==2.31.0
